requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
requests==2.32.3
httpx==0.28.1
aiofiles==24.1.0
python-dotenv==1.1.0
Pillow==11.2.1
playwright==1.53.0
groq
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
